Add simple assignments storage
